% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{make_windows_bins}
\alias{make_windows_bins}
\alias{make_windows_bins,GRANRepository-method}
\title{make_windows_bins
Return the logical that determines whether to build Windows binaries}
\usage{
make_windows_bins(repo)

\S4method{make_windows_bins}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
Logical indicating whether Windows binaries will be built
}
\description{
make_windows_bins
Return the logical that determines whether to build Windows binaries
}
