% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{metadatadir}
\alias{metadatadir}
\alias{metadatadir,GRANRepository-method}
\title{metadatadir
Return the full path to the pkg metadata directory for the final repository
deployment.}
\usage{
metadatadir(repo)

\S4method{metadatadir}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
The full path to the metadata directory
}
\description{
metadatadir
Return the full path to the pkg metadata directory for the final repository
deployment.
}
