% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{loadRepo}
\alias{loadRepo}
\alias{saveRepo}
\title{Backwards compatible load utility}
\usage{
loadRepo(filename)

saveRepo(repo, filename)
}
\arguments{
\item{filename}{The file to load}

\item{repo}{The GRANRepository object to save}
}
\description{
Load a repository serialized to an R code file

serialize a repository to a file so that it does not require GRANBase
to load
}
\examples{
repo = GRANRepository(GithubManifest("gmbecker/rpath"), basedir = tempdir())
fil = file.path(tempdir(), "repo.R")
saveRepo(repo, fil)
repo2 = loadRepo(fil)
}
