% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{RepoToList}
\alias{RepoToList}
\alias{RepoFromList}
\title{Transform a GRANRepository object into a list}
\usage{
RepoToList(repo)

RepoFromList(rlist)
}
\arguments{
\item{repo}{repository}

\item{rlist}{A list with entries that are slot name-value
pairs for a GRANRepository object}
}
\value{
a list suitable for use with RepoFromList

a GRANRepository object
}
\description{
Utility to transform a GRANRepository object into a list
so that repos saved using GRANBase can be loaded by GRAN
without requiring GRANBase
}
\examples{
repo = GRANRepository(GithubManifest("gmbecker/switchr"), basedir = tempdir())
lst = RepoToList(repo)
repo2 = RepoFromList(lst)
}
