% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updaterepo.R
\name{updateGRANRepoObject}
\alias{updateGRANRepoObject}
\title{updateGRANRepoObject}
\usage{
updateGRANRepoObject(object, ...)
}
\arguments{
\item{object}{The GRAN repo object}

\item{...}{Other parameters from RepoBuildParam}
}
\description{
Update the GRAN repo object with newer information. This is called
internally by loadRepo to ensure that the resulting repo doesn't
have stale logging closures and the like.
}
\note{
This function is cal not intended for use by the end user. This
    is called internally by loadRepo.
}
