% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{windowsbindir}
\alias{windowsbindir}
\alias{windowsbindir,GRANRepository-method}
\title{windowsbindir
Return the full path to the location of the windows binary builds}
\usage{
windowsbindir(repo)

\S4method{windowsbindir}{GRANRepository}(repo)
}
\arguments{
\item{repo}{a GRANRepository object}
}
\value{
Full path to the location of the windows binary builds
}
\description{
windowsbindir
Return the full path to the location of the windows binary builds
}
