% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdirs.R
\name{prepDirStructure}
\alias{prepDirStructure}
\title{prepDirStructure}
\usage{
prepDirStructure(basedir, repo_name, temp_repo, temp_checkout, tempLibLoc,
  destination, logfile, errlogfile)
}
\arguments{
\item{basedir}{The base directory. By default the temporary repository,
temporary install library, and package staging area will be located in
<basedir>/<subrepoName>/, while the  temporary source checkout will be in
the basedir itself.}

\item{repo_name}{The name of the repository, e.g. stable or devel}

\item{temp_repo}{Location to create the temporary repository}

\item{temp_checkout}{Location to create temporary checkouts/copies of package
source code}

\item{tempLibLoc}{Location to create the temporary installed package library
for use during the testing process}

\item{destination}{Base location (not including repository name) of the
final repository to be built.}

\item{logfile}{Comprehensive GRAN log file}

\item{errlogfile}{Error log file}
}
\description{
Initialize the directory structure for the GRAN repo
}
\note{
This function is not intended for use by the end user.
}
