% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGRANRepoObject.R
\name{updateGRANRepoObject}
\alias{updateGRANRepoObject}
\title{updateGRANRepoObject}
\usage{
updateGRANRepoObject(object, ...)
}
\arguments{
\item{object}{The GRAN repo object}

\item{...}{Other parameters from RepoBuildParam}
}
\description{
Update the GRAN repo object with newer information
}
